Refactor project to Vite with React 19 and Tailwind
