+++ lib/react-slidebox/src/components/slide-box.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import SlideBoxStyle from './slide-box.scss'
+import SlideBoxStyle from './slide-box.module.scss'
 import PropTypes from 'prop-types'
 
 const BannerList = (props) => {
