+++ lib/react-slidebox/src/components/slide-box.module.scss
