+++ package.json
@@ -4,33 +4,24 @@
     "description": "React Slidebox Test Project",
     "author": "Anapopo <favoursmoe@gmail.com>",
     "license": "MIT",
-    "private": false,
     "scripts": {
         "lib": "cd lib/react-slidebox && yarn build",
-        "build": "webpack",
+        "dev": "vite",
+        "build": "vite build",
+        "preview": "vite preview",
         "lint": "eslint src"
     },
-    "devDependencies": {
-        "@babel/core": "^7.1.2",
-        "@babel/plugin-proposal-class-properties": "^7.1.0",
-        "@babel/preset-env": "^7.1.0",
-        "@babel/preset-react": "^7.0.0",
-        "babel-loader": "^8.0.4",
-        "clean-webpack-plugin": "^0.1.19",
-        "css-loader": "^1.0.0",
-        "html-webpack-plugin": "^3.2.0",
-        "mini-css-extract-plugin": "^0.4.4",
-        "node-sass": "^4.9.4",
-        "optimize-css-assets-webpack-plugin": "^5.0.1",
-        "react-dom": "^16.6.0",
-        "sass-loader": "^7.1.0",
-        "style-loader": "^0.23.1",
-        "uglifyjs-webpack-plugin": "^2.0.1",
-        "webpack": "^4.23.1",
-        "webpack-cli": "^3.1.2"
-    },
     "dependencies": {
-        "prop-types": "^15.6.2",
-        "react": "^16.6.0"
+        "prop-types": "^15.8.1",
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0"
+    },
+    "devDependencies": {
+        "@vitejs/plugin-react-swc": "^3.6.0",
+        "autoprefixer": "^10.4.20",
+        "postcss": "^8.4.38",
+        "sass": "^1.77.8",
+        "tailwindcss": "^3.4.10",
+        "vite": "^5.4.1"
     }
 }
