--- public/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <title>React SlideBox</title>
-</head>
-<body>
-    <div id="root"></div>
-</body>
-</html>
